requirements.txt
# requirements.txt
aiohttp==3.9.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
pydantic==2.5.3
python-dotenv==1.0.0